Extract Utility Libraries into new package
