+++ orchestration-utils/build.gradle
@@ -14,23 +14,6 @@ java {
   withSourcesJar()
 }
 
-jar {
-  dependsOn(':parsing-utilities:jar')
-  dependsOn(':merlin-sdk:jar')
-  dependsOn(':merlin-driver:jar')
-  dependsOn(':constraints:jar')
-  dependsOn(':orchestration-utils:jar')
-
-  duplicatesStrategy = DuplicatesStrategy.EXCLUDE
-
-  manifest {
-    attributes('Main-Class': 'gov.nasa.jpl.aerie.stateless.Main')
-  }
-  from {
-    configurations.runtimeClasspath.collect { it.isDirectory() ? it : zipTree(it) }
-  }
-}
-
 test {
   useJUnitPlatform {
     includeEngines 'junit-jupiter'
@@ -49,23 +32,23 @@ jacocoTestReport {
   }
 }
 
+
+dependencies {
+  implementation project(':merlin-server')
+  implementation project(':merlin-driver')
+  implementation project(':parsing-utilities')
+  implementation 'commons-cli:commons-cli:1.8.0'
+
+  testImplementation 'org.junit.jupiter:junit-jupiter-engine:5.10.0'
+  testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
+}
+
+
 // Link references to standard Java classes to the official Java 11 documentation.
 javadoc.options.links 'https://docs.oracle.com/en/java/javase/11/docs/api/'
 javadoc.options.links 'https://commons.apache.org/proper/commons-lang/javadocs/api-3.9/'
 javadoc.options.addStringOption('Xdoclint:none', '-quiet')
 
-dependencies {
-  implementation project(':orchestration-utils')
-  implementation project(':merlin-driver')
-  implementation project(':merlin-server')
-  implementation 'commons-cli:commons-cli:1.8.0'
-
-  testImplementation 'org.junit.jupiter:junit-jupiter-engine:5.10.0'
-  testImplementation 'org.junit.jupiter:junit-jupiter-params:5.10.0'
-  testRuntimeOnly(project(':examples:foo-missionmodel'))
-
-  testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-}
 
 publishing {
   publications {
