+++ orchestration-utils/src/main/java/gov/nasa/jpl/aerie/orchestration/simulation/SimulationUtility.java
@@ -1,4 +1,4 @@
-package gov.nasa.jpl.aerie.stateless.simulation;
+package gov.nasa.jpl.aerie.orchestration.simulation;
 
 import gov.nasa.jpl.aerie.merlin.driver.DirectiveTypeRegistry;
 import gov.nasa.jpl.aerie.merlin.driver.MissionModel;
@@ -8,7 +8,6 @@ import gov.nasa.jpl.aerie.merlin.driver.SimulationDriver;
 import gov.nasa.jpl.aerie.merlin.driver.SimulationException;
 import gov.nasa.jpl.aerie.merlin.driver.SimulationResults;
 import gov.nasa.jpl.aerie.merlin.driver.resources.InMemorySimulationResourceManager;
-import gov.nasa.jpl.aerie.merlin.driver.resources.ResourceProfiles;
 import gov.nasa.jpl.aerie.merlin.driver.resources.SimulationResourceManager;
 import gov.nasa.jpl.aerie.merlin.driver.resources.StreamingSimulationResourceManager;
 import gov.nasa.jpl.aerie.merlin.protocol.model.ModelType;
@@ -46,7 +45,7 @@ public class SimulationUtility implements AutoCloseable {
    * Create a new SimulationUtility that manages resources using a StreamingSimulationResourceManager.
    * @param resourceStreamer a Consumer defining how the ResourceManager will stream resources.
    */
-  public SimulationUtility(Consumer<ResourceProfiles> resourceStreamer) {
+  public SimulationUtility(ResourceFileStreamer resourceStreamer) {
     this.exec = Executors.newSingleThreadExecutor();
     rmgr = new StreamingSimulationResourceManager(resourceStreamer);
   }
@@ -137,7 +136,7 @@ public class SimulationUtility implements AutoCloseable {
             plan.activityDirectives,
             plan.simulationStartTimestamp.toInstant(),
             simulationDuration,
-            plan.startTimestamp.toInstant(),
+            plan.planStartInstant(),
             plan.duration(),
             canceledListener,
             extentConsumer,
