+++ settings.gradle
@@ -42,3 +42,4 @@ include 'examples:config-without-defaults'
 include 'examples:minimal-mission-model'
 include 'examples:streamline-demo'
 include 'stateless-aerie'
+include 'orchestration-utils'
