+++ stateless-aerie/src/main/java/gov/nasa/jpl/aerie/stateless/Main.java
@@ -1,9 +1,10 @@
 package gov.nasa.jpl.aerie.stateless;
 
-import gov.nasa.jpl.aerie.stateless.simulation.CanceledListener;
-import gov.nasa.jpl.aerie.stateless.simulation.ResourceFileStreamer;
-import gov.nasa.jpl.aerie.stateless.simulation.SimulationExtentConsumer;
-import gov.nasa.jpl.aerie.stateless.simulation.SimulationResultsWriter;
+import gov.nasa.jpl.aerie.orchestration.simulation.CanceledListener;
+import gov.nasa.jpl.aerie.orchestration.PlanJsonParser;
+import gov.nasa.jpl.aerie.orchestration.simulation.ResourceFileStreamer;
+import gov.nasa.jpl.aerie.orchestration.simulation.SimulationExtentConsumer;
+import gov.nasa.jpl.aerie.orchestration.simulation.SimulationResultsWriter;
 import gov.nasa.jpl.aerie.merlin.driver.MissionModel;
 import gov.nasa.jpl.aerie.merlin.driver.MissionModelLoader;
 import gov.nasa.jpl.aerie.merlin.driver.SimulationException;
@@ -14,7 +15,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ExecutionException;
 
-import gov.nasa.jpl.aerie.stateless.simulation.SimulationUtility;
+import gov.nasa.jpl.aerie.orchestration.simulation.SimulationUtility;
 import org.apache.commons.cli.*;
 
 import javax.json.Json;
